Reorganize packages in a more convenient fashion
