+++ README.md
@@ -306,4 +306,4 @@ This work is published under the MIT license.
 Please see the `LICENSE` file for details.
 
 * * *
-Automatically generated by [autoreadme](https://github.com/jimmyfrasche/autoreadme) on 2019.05.19
+Automatically generated by [autoreadme](https://github.com/jimmyfrasche/autoreadme) on 2019.05.21
