+++ app/cli/cli.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"github.com/jawher/mow.cli"
 	"github.com/s-petit/birthday-pal/app"
-	"github.com/s-petit/birthday-pal/auth"
-	"github.com/s-petit/birthday-pal/email"
-	"github.com/s-petit/birthday-pal/remind"
-	"github.com/s-petit/birthday-pal/request"
+	"github.com/s-petit/birthday-pal/app/contact/auth"
+	"github.com/s-petit/birthday-pal/app/contact/request"
+	"github.com/s-petit/birthday-pal/app/email"
+	"github.com/s-petit/birthday-pal/app/remind"
 	"github.com/s-petit/birthday-pal/system"
 	"google.golang.org/api/people/v1"
 	"log"
