+++ app/contact/request/provider.go
@@ -1,7 +1,7 @@
 package request
 
 import (
-	"github.com/s-petit/birthday-pal/contact"
+	"github.com/s-petit/birthday-pal/app/contact"
 )
 
 //ContactsProvider calls a HTTP Contact API and returns a slice of Contacts regardless of protocol or authentication
