+++ app/contact/request/vcard/parser_test.go
@@ -2,7 +2,7 @@ package vcard
 
 import (
 	"github.com/mapaiva/vcard-go"
-	"github.com/s-petit/birthday-pal/contact"
+	"github.com/s-petit/birthday-pal/app/contact"
 	"github.com/s-petit/birthday-pal/testdata"
 	"github.com/stretchr/testify/assert"
 	"testing"
