+++ app/email/french_test.go
