+++ main.go
@@ -2,7 +2,7 @@ package main
 
 import (
 	"github.com/s-petit/birthday-pal/app"
-	"github.com/s-petit/birthday-pal/cli"
+	"github.com/s-petit/birthday-pal/app/cli"
 	"github.com/s-petit/birthday-pal/system"
 )
